Improve layout & Readability of labs/assignments
