+++ lab/refactoring/Refactoring_Composing_Methods/introduce_explaining_variable3.py
@@ -10,3 +10,6 @@ def is_cookeding_criteria_satisfied(time, temperature, pressure, desired_state):
     if desired_state == 'medium' and time * temperature * pressure * COOKED_CONSTANT >= MEDIUM:
         return True
     return False
+
+if __name__ == "__main__":
+    print(is_cookeding_criteria_satisfied(50, 450, 4500, "medium"))
