+++ lab/refactoring/Refactoring_Composing_Methods/replace_temp_with_query.py
@@ -45,7 +45,15 @@ class School:
         for e in top_employers:
             e.send(top_10_percent)
 
-students = [Student(2.1, 'Pinocchio'), Student(2.3, 'goku'), Student(2.7, 'toro'), 
-            Student(3.9, 'naruto'), Student(3.2,'kami'), Student(3,'guts')]
-school  = School(students)
-school.process_graduation()
+
+if __name__ == "__main__":
+    students = [
+        Student(2.1, 'Pinocchio'),
+        Student(2.3, 'goku'),
+        Student(2.7, 'toro'), 
+        Student(3.9, 'naruto'),
+        Student(3.2,'kami'),
+        Student(3,'guts')
+    ]
+    school  = School(students)
+    school.process_graduation()
