+++ lab/refactoring/Refactoring_Composing_Methods_2/rename_method.py
@@ -16,13 +16,14 @@ def get_value(li):  # TODO: Rename this function to reflect what it's doing.
             m = value
     return m
 
-
-li = [5, -1, 43, 32, 87, -100]
-print(get_value(li))
-
 ############################
 def process(sentence):  # TODO: Rename this function to reflect what it's doing.
     words = sentence[0:].split(' ')
     return words
 
-print(process('If you were a vegetable, you’d be a ‘cute-cumber.'))
+
+if __name__ == "__main__":
+    li = [5, -1, 43, 32, 87, -100]
+    print(get_value(li))
+
+    print(process('If you were a vegetable, you’d be a ‘cute-cumber.'))
