+++ lab/refactoring/Refactoring_Composing_Methods_2/split_temp_variable.py
@@ -1,0 +1,19 @@
+# by Kami Bigdely
+# Split temporary variable
+
+patty = 70  # [gr]
+pickle = 20  # [gr]
+tomatoes = 25  # [gr]
+lettuce = 15  # [gr]
+buns = 95  # [gr]
+kimchi = 30  # [gr]
+mayo = 5  # [gr]
+golden_fried_onion = 20  # [gr]
+
+if __name__ == "__main__":
+    sandwich_weight = (2 * patty + 4 * pickle + 3 *
+                       tomatoes + 2 * lettuce + 2 * buns)
+    print("NY Burger Weight", sandwich_weight)
+    sandwich_weight = (2 * patty + 4 * pickle + 3 * tomatoes
+                       + kimchi + mayo + golden_fried_onion + 2 * buns)
+    print("Seoul Kimchi Burger Weight", sandwich_weight)
