+++ lab/refactoring/Refactoring_Other_Refactoring/extract_superclass2.py
