+++ lab/refactoring/Refactoring_Simplifying_Conditional_Expressions/consolidate_conditional.py
