+++ lab/refactoring/Refactoring_Simplifying_Conditional_Expressions/consolidate_duplicate_conditional_fragments2.py
