+++ lab/refactoring/Refactoring_Simplifying_Conditional_Expressions/decompose_conditional.py
