+++ lab/refactoring/Refactoring_Simplifying_Conditional_Expressions/introduce_explaining_variable.py
