+++ lab/refactoring/Refactoring_Simplifying_Conditional_Expressions/remove_control_flag2.py
