+++ lab/refactoring/Refactoring_Simplifying_Conditional_Expressions/remove_control_flag3.py
@@ -12,4 +12,6 @@ def backdoor_srcmethod():
         else:
             print('doing other stuff.') 
 
-backdoor_srcmethod()
+
+if __name__ == "__main__":
+    backdoor_srcmethod()
