+++ lab/refactoring/Refactoring_Simplifying_Conditional_Expressions/replace_nested_conditional_with_gaurd_clauses2.py
@@ -1,5 +1,5 @@
 # Adapted from a Java code in the "Refactoring" book by Martin Fowler.
-# Replace nested conditional with gaurd clases.
+# Replace nested conditional with guard clases.
 ADJ_FACTOR = 0.7
 def get_adjusted_capital(capital, rate, duration, income):
     result = 0
@@ -9,5 +9,7 @@ def get_adjusted_capital(capital, rate, duration, income):
     
     return result
 
-adjusted_capital = get_adjusted_capital(50000, 4,10, 10000)
-print(adjusted_capital)
+
+if __name__ == "__main__":
+    adjusted_capital = get_adjusted_capital(50000, 4,10, 10000)
+    print(adjusted_capital)
